# src/main/java/ar/edu/utn/frc/tup/lciii/dtos/common/CountryDTOs/AmountOfCountries.java
package ar.edu.utn.frc.tup.lciii.dtos.common.CountryDTOs;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class AmountOfCountries {

    @JsonProperty("amountOfCountryToSave")
    private int amount;

}